Add message color print functionality
